requirements.txt
openai
google-generativeai
python-dotenv
pandas
pdf2image
pillow
requests
tqdm
